IOS-3883: Extract auxiliary types into separate source file
